refactor: Profile 컴포넌트 커스텀훅 분리
